Внес некоторые изменения (см.ниже): 1. В классе Buyer переименовал sale в saleValue 2. Переименовал класс Config в JDBCConnectionConfig 3. Добавил пакет dao.impl для оптимизации архитектуры 4. Добавил sql запрос на создание таблицы buyer (resources/migration)
