+++ src/main/java/config/JDBCConnectionConfig.java
@@ -5,7 +5,7 @@ import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.util.Properties;
 
-public class Config {
+public class JDBCConnectionConfig {
 
     private static String URL;
     private static String USERNAME;
