+++ src/main/java/entity/Buyer.java
@@ -1,14 +1,20 @@
-package entities;
+package entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-//Эдуард
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Buyer {
+
     private long id;
+
     private long cardNumber;
-    private long sale;
+
+    private long saleValue;
+
+    private Order order;
+
 }
