+++ src/main/java/entity/Product.java
@@ -1,4 +1,4 @@
-package entities;
+package entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
