+++ src/main/resources/migration/create-buyer-table.sql
@@ -1,0 +1,7 @@
+CREATE TABLE if NOT EXISTS buyer (
+    id SERIAL PRIMARY KEY,
+    card_number BIGINT NOT NULL,
+    sale_value BIGINT NOT NULL,
+
+    FOREIGN KEY (order_id) REFERENCES order(id)
+);
